Modularize Printer
I need to do several refactorings to adapt metap to the upcoming changes
to refinement and existential types, and I found that it is super hard
to do these refactorings when everything depends on everything.
