+++ semanticdb/metap/src/main/scala/scala/meta/internal/semanticdb3/Print.scala
@@ -3,7 +3,7 @@ package scala.meta.internal.semanticdb3
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
 import scala.meta.cli._
-import scala.meta.internal.metap.Printer
+import scala.meta.internal.metap.DocumentPrinter
 import scala.meta.metap.Settings
 
 object Print {
@@ -11,7 +11,7 @@ object Print {
   def document(doc: TextDocument): String = {
     val baos = new ByteArrayOutputStream()
     val ps = new PrintStream(baos)
-    val printer = new Printer(ps, doc)
+    val printer = new DocumentPrinter(ps, doc)
     printer.print()
     baos.toString()
   }
